# build: speed up webpack build using HardSource for caching