--- a/.gitignore
+++ b/.gitignore
@@ -29,3 +29,5 @@ npm-debug.log.*
 # Cypress testing. These will get downloaded as artifacts by buildkite.
 cypress/screenshots/
 cypress/vidoes/
+
+.cache/