--- a/app/app.scss
+++ b/app/app.scss
@@ -89,5 +89,5 @@ $theme-colors-light: (
 @import './components/symbols';
 @import './components/codeIntelStatusIndicator';
 @import 'node_modules/@sourcegraph/codeintellify/lib/HoverOverlay.scss';
-@import 'node_modules/@sourcegraph/react-loading-spinner/lib/LoadingSpinner.css';
+@import 'node_modules/@sourcegraph/react-loading-spinner/lib/LoadingSpinner';
 @import './extensions-client-common';