https://github.com/mzgoddard/hard-source-webpack-plugin

HardSource caches modules in `.cache/` and cuts down the build time:

- Initial build: 55s down to 15s 🚤💨 
- Incremental build: 10s down to 7s (based on light empirical evidence)

Here's what the output looks like 👀:

```
[hardsource:57550c93] Using 69 MB of disk space.
[hardsource:57550c93] Tracking node dependencies with: package-lock.json.
[hardsource:57550c93] Reading from cache 57550c93...
[16:36:53] › …  awaiting  Compiling...
...
```

Unfortunately, I ran into the same issue reported here ☹️ https://github.com/mzgoddard/hard-source-webpack-plugin/issues/416 and I avoided it by removing node-sass-import-once and dropping the `.css` extension on some SASS imports. What's left are some deprecation warnings that will be dropped soon https://github.com/sass/node-sass/issues/2362#issuecomment-401964451 These deprecation warnings can be ignored 🙈:

```
DEPRECATION WARNING on line 4, column 8 of /Users/chrismwendt/sourcegraph/browser-extensions/app/extensions-client-common.scss:
Including .css files with @import is non-standard behaviour which will be removed in future versions of LibSass.
Use a custom importer to maintain this behaviour. Check your implementations documentation on how to create a custom importer.
...
```