@chrismwendt 2018-08-20
I see, the removal of node-sass-import-once breaks the CSS:

![image](https://user-images.githubusercontent.com/1387653/44314850-694e5100-a3d2-11e8-9f0c-0696b873143c.png)

This shouldn't be a problem if/when we switch from SASS to some CSS-in-JS solution (e.g. aphrodite). Unless there's a way to fix this sooner, let's table HardSource until then.