@felixfbecker 2018-08-20
I don't know, I would not think so. I would recommend you post a comment on that thread saying that you identified this only happens with sass-import-once, because I don't see it mentioned in the thread.

The build of the browser-extension is not very optimized. I think there are other ways we can speed it up we can also try.